# perf: enable row ID bitset optimisation for equality predicates